+++ Warehouse_testing/WarehouseSolution.sln
@@ -3,7 +3,9 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.31605.320
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Warehouse_testing", "Warehouse_testing.csproj", "{402B8C66-4786-40CE-998D-6FCC1B14EA0A}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "WarehouseApp", "WarehouseApp.csproj", "{402B8C66-4786-40CE-998D-6FCC1B14EA0A}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WarehouseLibrary", "..\WarehouseLibrary\WarehouseLibrary.csproj", "{B4C09E25-57F4-4446-8CC8-12FEA2CFF427}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -15,6 +17,10 @@ Global
 		{402B8C66-4786-40CE-998D-6FCC1B14EA0A}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{402B8C66-4786-40CE-998D-6FCC1B14EA0A}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{402B8C66-4786-40CE-998D-6FCC1B14EA0A}.Release|Any CPU.Build.0 = Release|Any CPU
+		{B4C09E25-57F4-4446-8CC8-12FEA2CFF427}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B4C09E25-57F4-4446-8CC8-12FEA2CFF427}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B4C09E25-57F4-4446-8CC8-12FEA2CFF427}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B4C09E25-57F4-4446-8CC8-12FEA2CFF427}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
